Add conversion to float.
This fixes a compilation error in xcode:
"Invalid operands to binary expression ('basic_ostream<char, std::__1::char_traits<char> >' and 'c10::Scalar')"

+++ src/models/VGG16.hpp
@@ -37,7 +37,7 @@ public:
         auto maxId = torch::argmax(out).item();
         auto maxProb = torch::max(out).item();
 
-        std::cout << "Probability: " << maxProb << std::endl;
+        std::cout << "Probability: " << maxProb.toFloat() << std::endl;
         std::cout << "Label: " << mClassLabels[maxId.toInt()] << std::endl;
 
         return std::make_pair(maxId.toInt(), maxProb.toFloat());
